test: Move credentials tests to separate folder
